+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 
 * 1.17.1:
+    * Improved: COD gateway allows to select status which should be set to order. 
     * Fix: Scripts and compatibility with JRTO.
     * Fix: Notices, warnings and error on reports page.
 * 1.17 - 2015-04-16: